+++ src/model_config/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["nodes", "types"]
+__all__ = ["graphic_object", "model", "types"]