+++ src/model_config/types.py
@@ -52,6 +52,8 @@ class Constant:
         return "c"
 
 
+
+
 # any type
 class Any:
     def __init__(self, dt=None):